Adds a new `send_recv` method to the dk crate.
This puts the given data into a packet, along with a random device address, and then waits for a reply containing a matching device address. This allows the function to reject packets that were intended for another device, and makes it somewhat robust when multiple devices are running on the same channel at the same time.